# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Silvae.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5089/"
    }
  ],
  "paths": {
...
    "/api/daily-reports": {
      "get": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "SearchDailyReports",
        "parameters": [
          {
            "name": "JobOrderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "WorksiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "CrewUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyReportSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/daily-reports/export.csv": {
      "get": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "ExportDailyReportsCsv",
        "parameters": [
          {
            "name": "JobOrderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "WorksiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "CrewUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/daily-reports/export.pdf": {
      "get": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "ExportDailyReportsPdf",
        "parameters": [
          {
            "name": "JobOrderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "WorksiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "CrewUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/daily-reports/{reportId}": {
      "get": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "GetDailyReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/daily-reports/{reportId}/submit": {
      "post": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "SubmitDailyReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDailyReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/daily-reports/{reportId}/approve": {
      "post": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "ApproveDailyReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/daily-reports/{reportId}/reopen": {
      "post": {
        "tags": [
          "DailyReports"
        ],
        "operationId": "ReopenDailyReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/certifications": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "operationId": "GetCertifications",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "validOn",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Certifications"
        ],
        "operationId": "CreateCertification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCertificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/certifications/expiring": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "operationId": "GetExpiringCertifications",
        "parameters": [
          {
            "name": "withinDays",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/certifications/inspection": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "operationId": "GetCertificationInspection",
        "parameters": [
          {
            "name": "worksiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InspectionDayDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/certifications/{certificationId}": {
      "put": {
        "tags": [
          "Certifications"
        ],
        "operationId": "UpdateCertification",
        "parameters": [
          {
            "name": "certificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCertificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Certifications"
        ],
        "operationId": "DeleteCertification",
        "parameters": [
          {
            "name": "certificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CertificationDto": {
        "required": [
          "id",
          "userId",
          "displayName",
          "kind",
          "issuer",
          "validFrom",
          "expiresOn",
          "notes",
          "documentId",
          "isValidToday",
          "daysToExpiry",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "issuer": {
            "type": [
              "null",
              "string"
            ]
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "expiresOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "documentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isValidToday": {
            "type": "boolean"
          },
          "daysToExpiry": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DailyReportActivityDto": {
        "required": [
          "description",
          "quantity",
          "unit"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "unit": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DailyReportAuditDto": {
        "required": [
          "action",
          "actorId",
          "actorName",
          "version",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "actorName": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DailyReportCrewDto": {
        "required": [
          "userId",
          "displayName",
          "hours",
          "note"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "hours": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DailyReportDetailDto": {
        "required": [
          "id",
          "organizationId",
          "worksiteId",
          "worksiteCode",
          "worksiteName",
          "authorId",
          "authorName",
          "reportDate",
          "notes",
          "status",
          "totalHours",
          "signature",
          "signedBy",
          "signedByName",
          "signedAt",
          "version",
          "createdAt",
          "updatedAt",
          "crew",
          "activities",
          "safetyChecks",
          "photos",
          "audit"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteId": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteCode": {
            "type": "string"
          },
          "worksiteName": {
            "type": "string"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "reportDate": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "totalHours": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "signature": {
            "type": [
              "null",
              "string"
            ]
          },
          "signedBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "signedByName": {
            "type": [
              "null",
              "string"
            ]
          },
          "signedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "crew": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportCrewDto"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportActivityDto"
            }
          },
          "safetyChecks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportSafetyCheckDto"
            }
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportPhotoDto"
            }
          },
          "audit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportAuditDto"
            }
          }
        }
      },
      "DailyReportPhotoDto": {
        "required": [
          "localReference",
          "latitude",
          "longitude",
          "capturedAt",
          "caption"
        ],
        "type": "object",
        "properties": {
          "localReference": {
            "type": "string"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "caption": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DailyReportSafetyCheckDto": {
        "required": [
          "code",
          "isCompliant",
          "note"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "isCompliant": {
            "type": "boolean"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DailyReportSummaryDto": {
        "required": [
          "id",
          "worksiteId",
          "worksiteCode",
          "worksiteName",
          "jobOrderId",
          "jobOrderCode",
          "jobOrderName",
          "authorId",
          "authorName",
          "reportDate",
          "status",
          "totalHours",
          "crewCount",
          "photoCount",
          "hasSafetyFinding",
          "signature",
          "version",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteId": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteCode": {
            "type": "string"
          },
          "worksiteName": {
            "type": "string"
          },
          "jobOrderId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "jobOrderCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "jobOrderName": {
            "type": [
              "null",
              "string"
            ]
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "reportDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "totalHours": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "crewCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "photoCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasSafetyFinding": {
            "type": "boolean"
          },
          "signature": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InspectionDayDto": {
        "required": [
          "reportId",
          "reportDate",
          "worksiteId",
          "worksiteCode",
          "worksiteName",
          "status",
          "crew"
        ],
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string",
            "format": "uuid"
          },
          "reportDate": {
            "type": "string",
            "format": "date"
          },
          "worksiteId": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteCode": {
            "type": "string"
          },
          "worksiteName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "crew": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionPersonDto"
            }
          }
        }
      },
      "InspectionPersonDto": {
        "required": [
          "userId",
          "displayName",
          "hours",
          "certifications"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "hours": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionCertificationDto"
            }
          }
        }
      },
...
      "SubmitDailyReportRequest": {
        "required": [
          "signature"
        ],
        "type": "object",
        "properties": {
          "signature": {
            "type": "string"
          }
        }
      },
...
      "UpsertCertificationRequest": {
        "required": [
          "userId",
          "kind",
          "validFrom",
          "expiresOn",
          "issuer",
          "notes",
          "documentId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "expiresOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "issuer": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "documentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}